[ios] Reintroduced static library for iOS 7
iOS tests are still run against the dynamic framework on iOS 8, however. Adjusted +[MGLMapView resourceImageNamed:] to look for a Mapbox.bundle if built as a static library. Removed the faux framework target. Copyedited the podspec description because you keep using that word “Cocoa bindings” but I do not think it means what you think it means.